Major rehaul of precog algorithm
The math I was doing for deciding if sufficient molecule
variants had been seen to declare that none were assumed
was over-complicated and incorrect (thank you ToughThought
for catching this). Did a major rethink, and did away with
the calculation of minimum or maximum run counts to perform
variant checks for.
Instead, continue running until all variants have been seen.

To decide if a variant's failure was due to precog, stop if
we ever see any variant fail X times; where X is based on
the solution's observed success rate.

The math to handle impossible variants was unnecessary;
the exact and only variants that are impossible are the same
as the assumed first molecule, if and only if it was unique
to its bucket; ignore these variants explicitly as needed.

To address the issue of it potentially taking many runs to see
all variants (previously handled by stopping after doing naive
probabilistic calculations of how many runs was 'enough'),
implement seed-skipping; runs are performed as normal until
the success rate is calculated to sufficient confidence
(to ensure we don't bias the success rate measurement), but after
that, any seeds that consist entirely of molecule variants that
we have seen succeed before are skipped.
This ensures that the runtime is bounded, as every run either
gets closer to succeeding on all variants or gets its variants
closer to failing the X threshold.